TypicalAddressBook: Add a layer of abstraction
Before this, addressbook was only tested with persons. Since we are also storing accounts,
let's have a class that can properly generate a typical addressbook that contains both
persons and accounts in it.

Let's do that by creating a new java class called TypicalAddressBook which will be responsible of
populating the addressbook with the respective Typical<Object>.
