+++ src/test/data/XmlSerializableAddressBookTest/typicalPersonsAddressBook.xml
@@ -55,4 +55,16 @@
         <address>4th street</address>
         <remark></remark>
     </persons>
+    <accounts>
+        <username>azhikai</username>
+        <password>1122qq!@#</password>
+    </accounts>
+  <accounts>
+    <username>demo1</username>
+    <password>1122</password>
+  </accounts>
+  <accounts>
+    <username>demo2</username>
+    <password>22qq</password>
+  </accounts>
 </addressbook>
